// front-end/src/modules/admin/Users/users.types.ts
export interface IUserListDetail {
  email_id: string;
  first_name: string;
  id: string;
  last_name: string;
  user_id: string;
  user_role: string;
  user_status: string;
  user_phone_no: string;
  user_address: string;
  occupation: string;
  is_admin_contacted: boolean;
  position_title: string;
}